+++ Source/AllOverIt/Extensions/EnumerableExtensions.Parallel.cs
@@ -22,7 +22,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of tasks that will be created.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsTaskAsync<TType>(this IEnumerable<TType> items, Func<TType, Task> func, int degreeOfParallelism,
+        public static Task ForEachAsTaskAsync<TType>(this IEnumerable<TType> items, Func<TType, CancellationToken, Task> func, int degreeOfParallelism,
             CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -52,7 +52,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of tasks that will be created.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsTaskAsync<TType, TInput>(this IEnumerable<TType> items, Func<TType, TInput, Task> func, TInput input,
+        public static Task ForEachAsTaskAsync<TType, TInput>(this IEnumerable<TType> items, Func<TType, TInput, CancellationToken, Task> func, TInput input,
             int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -84,7 +84,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of tasks that will be created.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsTaskAsync<TType, TInput1, TInput2>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, Task> func, 
+        public static Task ForEachAsTaskAsync<TType, TInput1, TInput2>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, CancellationToken, Task> func, 
             TInput1 input1, TInput2 input2, int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -118,7 +118,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of tasks that will be created.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsTaskAsync<TType, TInput1, TInput2, TInput3>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, TInput3, Task> func,
+        public static Task ForEachAsTaskAsync<TType, TInput1, TInput2, TInput3>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, TInput3, CancellationToken, Task> func,
             TInput1 input1, TInput2 input2, TInput3 input3, int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -155,7 +155,7 @@ namespace AllOverIt.Extensions
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
         public static Task ForEachAsTaskAsync<TType, TInput1, TInput2, TInput3, TInput4>(this IEnumerable<TType> items,
-            Func<TType, TInput1, TInput2, TInput3, TInput4, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
+            Func<TType, TInput1, TInput2, TInput3, TInput4, CancellationToken, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
             int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -187,7 +187,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of parallel operations.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsParallelAsync<TType>(this IEnumerable<TType> items, Func<TType, Task> func, int degreeOfParallelism,
+        public static Task ForEachAsParallelAsync<TType>(this IEnumerable<TType> items, Func<TType, CancellationToken, Task> func, int degreeOfParallelism,
             CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -212,7 +212,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of parallel operations.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsParallelAsync<TType, TInput>(this IEnumerable<TType> items, Func<TType, TInput, Task> func, TInput input, int degreeOfParallelism,
+        public static Task ForEachAsParallelAsync<TType, TInput>(this IEnumerable<TType> items, Func<TType, TInput, CancellationToken, Task> func, TInput input, int degreeOfParallelism,
             CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -239,7 +239,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of parallel operations.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsParallelAsync<TType, TInput1, TInput2>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, Task> func,
+        public static Task ForEachAsParallelAsync<TType, TInput1, TInput2>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, CancellationToken, Task> func,
             TInput1 input1, TInput2 input2, int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -268,7 +268,7 @@ namespace AllOverIt.Extensions
         /// <param name="degreeOfParallelism">Determines the maximum number of parallel operations.</param>
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
-        public static Task ForEachAsParallelAsync<TType, TInput1, TInput2, TInput3>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, TInput3, Task> func,
+        public static Task ForEachAsParallelAsync<TType, TInput1, TInput2, TInput3>(this IEnumerable<TType> items, Func<TType, TInput1, TInput2, TInput3, CancellationToken, Task> func,
             TInput1 input1, TInput2 input2, TInput3 input3, int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -300,7 +300,7 @@ namespace AllOverIt.Extensions
         /// <param name="cancellationToken">A cancellation token that can cancel the processing.</param>
         /// <returns>A task that will complete when all items have been processed.</returns>
         public static Task ForEachAsParallelAsync<TType, TInput1, TInput2, TInput3, TInput4>(this IEnumerable<TType> items,
-            Func<TType, TInput1, TInput2, TInput3, TInput4, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
+            Func<TType, TInput1, TInput2, TInput3, TInput4, CancellationToken, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
             int degreeOfParallelism, CancellationToken cancellationToken = default)
         {
             _ = items.WhenNotNull(nameof(items));
@@ -324,19 +324,19 @@ namespace AllOverIt.Extensions
 
         #region ProcessPartitionAsync
 
-        private static async Task ProcessPartitionAsync<TType>(IEnumerator<TType> partition, Func<TType, Task> func, CancellationToken cancellationToken)
+        private static async Task ProcessPartitionAsync<TType>(IEnumerator<TType> partition, Func<TType, CancellationToken, Task> func, CancellationToken cancellationToken)
         {
             using (partition)
             {
                 while (partition.MoveNext())
                 {
                     cancellationToken.ThrowIfCancellationRequested();
-                    await func.Invoke(partition.Current).ConfigureAwait(false);
+                    await func.Invoke(partition.Current, cancellationToken).ConfigureAwait(false);
                 }
             }
         }
 
-        private static async Task ProcessPartitionAsync<TType, TInput>(IEnumerator<TType> partition, Func<TType, TInput, Task> func, TInput input,
+        private static async Task ProcessPartitionAsync<TType, TInput>(IEnumerator<TType> partition, Func<TType, TInput, CancellationToken, Task> func, TInput input,
             CancellationToken cancellationToken)
         {
             using (partition)
@@ -344,12 +344,12 @@ namespace AllOverIt.Extensions
                 while (partition.MoveNext())
                 {
                     cancellationToken.ThrowIfCancellationRequested();
-                    await func.Invoke(partition.Current, input).ConfigureAwait(false);
+                    await func.Invoke(partition.Current, input, cancellationToken).ConfigureAwait(false);
                 }
             }
         }
 
-        private static async Task ProcessPartitionAsync<TType, TInput1, TInput2>(IEnumerator<TType> partition, Func<TType, TInput1, TInput2, Task> func,
+        private static async Task ProcessPartitionAsync<TType, TInput1, TInput2>(IEnumerator<TType> partition, Func<TType, TInput1, TInput2, CancellationToken, Task> func,
             TInput1 input1, TInput2 input2, CancellationToken cancellationToken)
         {
             using (partition)
@@ -357,12 +357,12 @@ namespace AllOverIt.Extensions
                 while (partition.MoveNext())
                 {
                     cancellationToken.ThrowIfCancellationRequested();
-                    await func.Invoke(partition.Current, input1, input2).ConfigureAwait(false);
+                    await func.Invoke(partition.Current, input1, input2, cancellationToken).ConfigureAwait(false);
                 }
             }
         }
 
-        private static async Task ProcessPartitionAsync<TType, TInput1, TInput2, TInput3>(IEnumerator<TType> partition, Func<TType, TInput1, TInput2, TInput3, Task> func,
+        private static async Task ProcessPartitionAsync<TType, TInput1, TInput2, TInput3>(IEnumerator<TType> partition, Func<TType, TInput1, TInput2, TInput3, CancellationToken, Task> func,
             TInput1 input1, TInput2 input2, TInput3 input3, CancellationToken cancellationToken)
         {
             using (partition)
@@ -370,13 +370,13 @@ namespace AllOverIt.Extensions
                 while (partition.MoveNext())
                 {
                     cancellationToken.ThrowIfCancellationRequested();
-                    await func.Invoke(partition.Current, input1, input2, input3).ConfigureAwait(false);
+                    await func.Invoke(partition.Current, input1, input2, input3, cancellationToken).ConfigureAwait(false);
                 }
             }
         }
 
         private static async Task ProcessPartitionAsync<TType, TInput1, TInput2, TInput3, TInput4>(IEnumerator<TType> partition,
-            Func<TType, TInput1, TInput2, TInput3, TInput4, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
+            Func<TType, TInput1, TInput2, TInput3, TInput4, CancellationToken, Task> func, TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4,
             CancellationToken cancellationToken)
         {
             using (partition)
@@ -384,7 +384,7 @@ namespace AllOverIt.Extensions
                 while (partition.MoveNext())
                 {
                     cancellationToken.ThrowIfCancellationRequested();
-                    await func.Invoke(partition.Current, input1, input2, input3, input4).ConfigureAwait(false);
+                    await func.Invoke(partition.Current, input1, input2, input3, input4, cancellationToken).ConfigureAwait(false);
                 }
             }
         }